requirements.txt
django==4.2.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
psycopg2==2.9.6
djangorestframework-simplejwt==5.2.2
django-rest-auth==0.9.5
pillow==9.5.0
